+++ tests/text/test_bleu.py
@@ -19,34 +19,9 @@ from nltk.translate.bleu_score import SmoothingFunction, corpus_bleu
 from torch import tensor
 
 from tests.text.helpers import INPUT_ORDER, TextTester
+from tests.text.inputs import _inputs_multiple_references
 from torchmetrics.functional.text.bleu import bleu_score
 from torchmetrics.text.bleu import BLEUScore
-
-# example taken from
-# https://www.nltk.org/api/nltk.translate.html?highlight=bleu%20score#nltk.translate.bleu_score.corpus_bleu
-# EXAMPLE 1
-HYPOTHESIS_A = "It is a guide to action which ensures that the military always obeys the commands of the party"
-REFERENCE_1A = "It is a guide to action that ensures that the military will forever heed Party commands"
-REFERENCE_2A = "It is a guiding principle which makes the military forces always being under the command of the Party"
-REFERENCE_3A = "It is the practical guide for the army always to heed the directions of the party"
-
-# EXAMPLE 2
-HYPOTHESIS_B = "he read the book because he was interested in world history"
-REFERENCE_1B = "he was interested in world history because he read the book"
-
-# EXAMPLE 3
-HYPOTHESIS_C = "the cat the cat on the mat"
-REFERENCE_1C = "the cat is on the mat"
-REFERENCE_2C = "there is a cat on the mat"
-
-TUPLE_OF_REFERENCES = (
-    ((REFERENCE_1A, REFERENCE_2A, REFERENCE_3A), tuple([REFERENCE_1B])),
-    (tuple([REFERENCE_1B]), (REFERENCE_1C, REFERENCE_2C)),
-    (REFERENCE_1B),
-)
-TUPLE_OF_HYPOTHESES = ((HYPOTHESIS_A, HYPOTHESIS_B), (HYPOTHESIS_B, HYPOTHESIS_C), (HYPOTHESIS_B))
-
-BATCHES = {"preds": TUPLE_OF_HYPOTHESES, "targets": TUPLE_OF_REFERENCES}
 
 # https://www.nltk.org/api/nltk.translate.html?highlight=bleu%20score#nltk.translate.bleu_score.SmoothingFunction
 smooth_func = SmoothingFunction().method2
@@ -75,7 +50,7 @@ def _compute_bleu_metric_nltk(list_of_references, hypotheses, weights, smoothing
 )
 @pytest.mark.parametrize(
     ["preds", "targets"],
-    [(BATCHES["preds"], BATCHES["targets"])],
+    [(_inputs_multiple_references.preds, _inputs_multiple_references.targets)],
 )
 class TestBLEUScore(TextTester):
     @pytest.mark.parametrize("ddp", [False, True])