+++ tests/text/test_chrf.py
@@ -5,37 +5,13 @@ import pytest
 from torch import Tensor, tensor
 
 from tests.text.helpers import INPUT_ORDER, TextTester
+from tests.text.inputs import _inputs_multiple_references, _inputs_single_sentence_multiple_references
 from torchmetrics.functional.text.chrf import chrf_score
 from torchmetrics.text.chrf import CHRFScore
 from torchmetrics.utilities.imports import _SACREBLEU_AVAILABLE
 
 if _SACREBLEU_AVAILABLE:
     from sacrebleu.metrics import CHRF
-
-# example taken from
-# https://www.nltk.org/api/nltk.translate.html?highlight=bleu%20score#nltk.translate.bleu_score.corpus_bleu and adjusted
-# EXAMPLE 1
-HYPOTHESIS_A = "It is a guide to action which ensures that the military always obeys the commands of the party"
-REFERENCE_1A = "It is a guide to action that ensures that the military will forever heed Party commands"
-REFERENCE_2A = "It is a guiding principle which makes the military forces always being under the command of the Party"
-
-# EXAMPLE 2
-HYPOTHESIS_B = "he read the book because he was interested in world history"
-REFERENCE_1B = "he was interested in world history because he read the book"
-REFERENCE_2B = "It is the practical guide for the army always to heed the directions of the party"
-
-# EXAMPLE 3 (add intentionally whitespaces)
-HYPOTHESIS_C = "the cat the   cat on the mat "
-REFERENCE_1C = "the  cat is     on the mat "
-REFERENCE_2C = "there is a   cat on the mat"
-
-TUPLE_OF_REFERENCES = (
-    ((REFERENCE_1A, REFERENCE_2A), (REFERENCE_1B, REFERENCE_2B)),
-    ((REFERENCE_1B, REFERENCE_2B), (REFERENCE_1C, REFERENCE_2C)),
-)
-TUPLE_OF_HYPOTHESES = ((HYPOTHESIS_A, HYPOTHESIS_B), (HYPOTHESIS_B, HYPOTHESIS_C))
-
-BATCHES = {"preds": TUPLE_OF_HYPOTHESES, "targets": TUPLE_OF_REFERENCES}
 
 
 def sacrebleu_chrf_fn(
@@ -68,7 +44,7 @@ def sacrebleu_chrf_fn(
 )
 @pytest.mark.parametrize(
     ["preds", "targets"],
-    [(BATCHES["preds"], BATCHES["targets"])],
+    [(_inputs_multiple_references.preds, _inputs_multiple_references.targets)],
 )
 @pytest.mark.skipif(not _SACREBLEU_AVAILABLE, reason="test requires sacrebleu")
 class TestCHRFScore(TextTester):
@@ -150,15 +126,15 @@ def test_chrf_empty_class():
 
 
 def test_chrf_return_sentence_level_score_functional():
-    hyp = [HYPOTHESIS_B]
-    ref = [[REFERENCE_1B, REFERENCE_2B]]
+    hyp = _inputs_single_sentence_multiple_references.preds
+    ref = _inputs_single_sentence_multiple_references.targets
     _, chrf_sentence_score = chrf_score(ref, hyp, return_sentence_level_score=True)
     isinstance(chrf_sentence_score, Tensor)
 
 
 def test_chrf_return_sentence_level_class():
     chrf = CHRFScore(return_sentence_level_score=True)
-    hyp = [HYPOTHESIS_B]
-    ref = [[REFERENCE_1B, REFERENCE_2B]]
+    hyp = _inputs_single_sentence_multiple_references.preds
+    ref = _inputs_single_sentence_multiple_references.targets
     _, chrf_sentence_score = chrf(ref, hyp)
     isinstance(chrf_sentence_score, Tensor)