+++ tests/text/test_mer.py
@@ -3,6 +3,7 @@ from typing import Callable, List, Union
 import pytest
 
 from tests.text.helpers import INPUT_ORDER, TextTester
+from tests.text.inputs import _inputs_error_rate_batch_size_1, _inputs_error_rate_batch_size_2
 from torchmetrics.utilities.imports import _JIWER_AVAILABLE
 
 if _JIWER_AVAILABLE:
@@ -13,19 +14,6 @@ else:
 from torchmetrics.functional.text.mer import match_error_rate
 from torchmetrics.text.mer import MatchErrorRate
 
-BATCHES_1 = {"preds": [["hello world"], ["what a day"]], "targets": [["hello world"], ["what a wonderful day"]]}
-
-BATCHES_2 = {
-    "preds": [
-        ["i like python", "what you mean or swallow"],
-        ["hello duck", "i like python"],
-    ],
-    "targets": [
-        ["i like monthy python", "what do you mean, african or european swallow"],
-        ["hello world", "i like monthy python"],
-    ],
-}
-
 
 def _compute_mer_metric_jiwer(prediction: Union[str, List[str]], reference: Union[str, List[str]]):
     return compute_measures(reference, prediction)["mer"]
@@ -35,8 +23,8 @@ def _compute_mer_metric_jiwer(prediction: Union[str, List[str]], reference: Unio
 @pytest.mark.parametrize(
     ["preds", "targets"],
     [
-        (BATCHES_1["preds"], BATCHES_1["targets"]),
-        (BATCHES_2["preds"], BATCHES_2["targets"]),
+        (_inputs_error_rate_batch_size_1.preds, _inputs_error_rate_batch_size_1.targets),
+        (_inputs_error_rate_batch_size_2.preds, _inputs_error_rate_batch_size_2.targets),
     ],
 )
 class TestMatchErrorRate(TextTester):