+++ torchmetrics/detection/map.py
@@ -371,7 +371,9 @@ class MAP(Metric):
         ignore_area = (areas < area_range[0]) | (areas > area_range[1])
 
         # sort dt highest score first, sort gt ignore last
-        ignore_area_sorted, gtind = torch.sort(ignore_area)
+        ignore_area_sorted, gtind = torch.sort(ignore_area.to(torch.uint8))
+        # Convert to uint8 temporarily and back to bool, because "Sort currently does not support bool dtype on CUDA"
+        ignore_area_sorted = ignore_area_sorted.to(torch.bool)
         gt = gt[gtind]
         scores = self.detection_scores[id]
         scores_filtered = scores[det_label_mask]
@@ -393,7 +395,7 @@ class MAP(Metric):
             for idx_iou, t in enumerate(self.iou_thresholds):
                 for idx_det in range(nb_det):
                     m = MAP._find_best_gt_match(t, nb_gt, gt_matches, idx_iou, gt_ignore, ious, idx_det)
-                    if m is not -1:
+                    if m != -1:
                         det_ignore[idx_iou, idx_det] = gt_ignore[m]
                         det_matches[idx_iou, idx_det] = True
                         gt_matches[idx_iou, m] = True